Rename the Ruby gem to spreen-wiki
spreen — a blend of the falcon's stoop and the preen that follows, after
Hayato (隼人, falcon man). The gem becomes spreen-wiki with the code under
Spreen::Wiki (lib/spreen/wiki plus a lib/spreen-wiki.rb require shim), the
executable becomes spreen and the config file .spreen.yml; RBS signatures
are regenerated for the new layout.
