+++ ruby/exe/spreen
@@ -1,0 +1,6 @@
+#!/usr/bin/env ruby
+# frozen_string_literal: true
+
+require 'spreen/wiki'
+
+exit Spreen::Wiki::CLI.start
