--- ruby/exe/wiki-organise
@@ -1,6 +1,0 @@
-#!/usr/bin/env ruby
-# frozen_string_literal: true
-
-require 'github_wiki_organiser'
-
-exit GithubWikiOrganiser::CLI.start
