+++ ruby/lib/spreen-wiki.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+# rbs_inline: enabled
+
+# Convenience shim so that both `require 'spreen-wiki'` (the gem name) and
+# `require 'spreen/wiki'` (the conventional path) work.
+require_relative 'spreen/wiki'
