+++ ruby/lib/spreen/wiki/templates/category/japanese.md
