+++ ruby/lib/spreen/wiki/templates/owner/japanese.md
