+++ ruby/test/home_test.rb
@@ -2,12 +2,12 @@
 # rbs_inline: enabled
 
 require_relative 'application_test'
-require_relative '../lib/github_wiki_organiser/home'
+require_relative '../lib/spreen/wiki/home'
 
 class HomeTest < ApplicationTest
   def setup(group_by: 'Owner', language: 'English', home_overflow: false)
     super
-    GithubWikiOrganiser::Home.run(base_path:, group_by:, language:, home_overflow:, organisation: 'test-org')
+    Spreen::Wiki::Home.run(base_path:, group_by:, language:, home_overflow:, organisation: 'test-org')
     @path_to_home           = File.join(base_path, 'Home.md')
     @home                   = File.read(path_to_home)
     @path_to_wikis_by_owner = File.join(base_path, 'wikis-by-owner')
